services/kafka-producer-python/requirements.txt
kafka-python
yfinance
pandas
